Adding network stuff

+++ src/SystemLayer/Windows/SystemLayerWindows_IO.cpp
@@ -101,7 +101,7 @@ namespace nl::systemlayer::defaults
         while (remaining != 0)
         {
             DWORD dw;
-            if (!ReadFile(hFile, ptr, remaining, &dw, nullptr) ||
+            if (!ReadFile(hFile, ptr, (DWORD)remaining, &dw, nullptr) ||
                 dw == 0)
                 break;
 
@@ -120,7 +120,7 @@ namespace nl::systemlayer::defaults
         while (remaining != 0)
         {
             DWORD dw;
-            if (!WriteFile(hFile, ptr, remaining, &dw, nullptr) ||
+            if (!WriteFile(hFile, ptr, (DWORD)remaining, &dw, nullptr) ||
                 dw == 0)
                 break;
